Accept RIOs with prefix length in the closed interval [48, 64]
This change adds the appropriate initialization code and constants to
accept what is believed to be a reasonable range of RIOs. Prior to this
change, only RIOs with prefix length of zero were accepted.

RIO bounds are set in such a way that we fail closed, that is, we only
start accepting RIOs if we've successfully applied the limits [48, 64].

Bug: 33333670
Test: as follows
  - built (angler)
  - flashed
  - booted
  - runtest -x netd_integration_test.cpp passed
  - runtest.py frameworks-net passed
  - runtest.py frameworks-wifi passed

Change-Id: I35497f5d2c0a23069bf7b000ccbdd7aad5709ef3