Fix ABI incompatibility for Netlink XFRM on Fugu
Fugu is compiled with an x86 userspace and an x86_64
kernel. This means that there is no guarantee of ABI
compatibility between the kernel and userspace. The
xfrm_usersa_info struct is one such place where the
compatibility happens to not exist due to struct
alignment differences. This CL patches the
xfrm_usersa_info struct to match the kernel's 64-bit
alignment in at least the case of x86 vs x86_64.

Bug: 37252170
Test: CTS - IpSecManagerTest passes
Merged-In: Ic08a75d543f92f7fa5e0cf8b4277de12464fd406
Change-Id: Ic08a75d543f92f7fa5e0cf8b4277de12464fd406