Always return a meaningful error from setIpv6AddrGenMode.
SetProperty does not always set errno to a meaningful value. This
is confusing, particularly in the case where the errno happens to
be set to EOPNOTSUPP: in that case IpManager will interpret the
error as being "kernel does not support stable privacy addresses"
and continue, instead of failing as it should.

Bug: 17613910
Test: bullhead builds, boots, connects to wifi
Change-Id: I058f5fe2dd1085036d669d89a4cac0b3b0c0949c