Fix a memory leak found by the static analyzer.
The pthread_* functions never set errno; they return EINVAL (and
others) directly.

The leak the analyzer was complaining about was in the condition after
scoped_attr.detach(). If errno was 0, we'd consider that a success in
DnsProxyListener.cpp, and proceed to free nothing.

That said, this line should probably never fail in the first place...

Bug: 27101951
Test: mma. Warnings about memory leak are gone.
Change-Id: Iebae756b9fd3001c0c41c2843dffcc64e52f4d05