Improve the Random Seeding of SPI Generator
Re-seeing the SPI generator with time_t caused
initial seeds generated at approximately the
same time to be the same. This means that SPIs
were either the same if the xfrm_id differed,
or were sequential if the xfrm_id was the same.
While that's technically not broken,
it's not good behavior. This patch converts to
using a better random seed, which in turn improves
the entropy of the generated SPIs.

Bug: 67507803
Test: manual
Change-Id: I58b84dbca17cfa6c27534de15416bd8906195739