WakeupController: remove --nflog-range
Trying to use the --nflog-range option for truncating packets received
with NFLOG does actually nothing and causes warnings to be printed by
IptablesRestoreController in netd.

The warning was added in https://patchwork.kernel.org/patch/9191269/
to iptables which also introduces a new XT_NFLOG_F_COPY_LEN
constant for passing down the packet copy len argument to the kernel
via a new --nflog-size option in iptables.

This patch removes the --nflog-range which is not necessary.
In reality WakeupController already specify a range argument when
subcribing to NFLOG using the NFLogListener interface in netd.
This sets the NFULNL_COPY_PACKET argument when doing the initial
Netlink request to NFLOG.

Bug: 73918721
Test: built, flashed, checked wakeup events are still correctly logged,
      checked that warnings are not printed anymore.

Change-Id: I742aaecf8563d5e476805301e0435a412bb37e6a