Restart main when netd restarts
Because netd crashes very rarely, and because it
maintains some state along with the system server,
and because upon restart netd starts first, vendors
who listen to netd restarts may begin recovery
procedures prior to the system server restarting if
it crashes, we need to ensure that if netd crashes
and restarts, the system server also restarts. This
means that now regardless of which goes down, the
restart is always the same (rather than asymmetrical
as it was before).

Bug: 74560705
Test: manual (crashed system server, crashed netd)
Change-Id: I27ecdc59a92f1fa0703e372e987e80482ad6f7f3