Add Tunnel Interface Prefix in INetd
Declare the ipsec tunnel interface prefix
as a constant in the NETD AIDL. This allows
the constant to be, well, constant between
the Java and native sides of the binder. Since
it is shared information, this provides an
added level of safety.

Note, this is being done as a follow-up to a
CL which created this cross-process dependency
on the name.

Bug: 74560705
Test: atest netd_integration_test;
      atest FrameworksNetTest
Change-Id: Ibc66b014692a978ebfcc7856dcc26524618668a8
(cherry picked from commit 1081b18a08fdbf19cb43cb09b2e2cfd05647e32a)