Do not count dropped packet on egress side
In the configuration of firewall controller and bandwidth controller, we
count inbound packets even if we later drop them, but we count outbound
packets only if we don't drop them. The eBPF program should be
consistent with that configuration. Also, we should always use the
socket uid to do the owner match.

Bug: 77990419
Bug: 30950746
Test: bpf program load when device boot
Change-Id: If5a6ca21fb76001d28190c5996a14391cbfff0a2