Exempt Unsupported Devs from XfrmControllerInit
For devices that do not support IPsec, or for those
devices which run a 64-bit kernel with 32-bit
userspace, IPsec features cannot be enabled. As other
tests ensure that for devices which support IPsec,
the feature is enabled in the kernel, allow them to
pass VTS tests without assuming they support IPsec.

Bug: 62778032
Test: TestXfrmControllerInit on AOSP/master - marlin
Change-Id: I18fad95646e2a9d7632ca13480d7d0ce13dafaa2