Add xfrm_userpolicy_info to netd compat structs
Netd has a 32-bit compatibility build flag for 32-bit userspace with a
64-bit kernel where some xfrm structs have different alignments. The
list of structs was missing xfrm_userpolicy_info which is used to set
the policy for a socket. The end result was that transformations were
not correctly applied for sockets as the kernel expected more data than
netd provided.

This adds xfrm_userpolicy_info to the list of compatibility structures
as well as static asserts to ensure that builds will fail if the
assumptions about the struct layout changes.

BUG: 80319713
Test: run cts -m CtsNetTestCases on 32-bit device with 64-bit kernel
Change-Id: I73d157db6beb1773c75c02431f50a05a55ef9c46