Remove bpf map iterator class
The bpf map iterator class implementation is not very helpful since the
map iterate process may fail because of the bpf syscall. So using a
iterator to go through the map is not very robust if it cannot through
exceptions when it failed. Instead, the map class will have some iterate
function that go through the map and apply operations based on the
lambda function passed in. And it can return failure when the
actual bpf syscall failed.

Bug: 78250686
Test: ./libbpf_test

Change-Id: Ie431ebd6e46c228347c3852fb5b3556534088439