Merge all uid owner match map into one
The bpf maps used for storing the uid owner match information are using
the same key value pairs and the duplication can be reduced by using one
single value to store all the match information for a given uid and use
a configuration map to store the chain that is currently enabled. This
migration can save some kernel memory space and simplify the owner match
process.

Bug: 79781072
Test: netd_unit_test,
      com.android.cts.net.HostsideRestrictBackgroundNetworkTests

Change-Id: I9658321e9d4a87eaa724231d33a474113dd75019